+++ web/src/hooks/analyze/useCardNavigation.ts
