+++ web/src/hooks/auth/useSignUpTerms.ts
