+++ web/src/hooks/community/mutations/useDeletePostMutation.ts
